Fix V3 and V3 switching
